third_party/blink/web_tests/platform/mac-mac14-arm64/external/wpt/webrtc-encoded-transform/script-transform-generateKeyFrame-simulcast.https-expected.txt

This is a testharness.js-based test.
[FAIL] generateKeyFrame works with simulcast rids
  assert_equals: expected "success" but got "failure"
[FAIL] generateKeyFrame for rid that was negotiated away fails
  assert_equals: expected "success" but got "failure"
[FAIL] generateKeyFrame with rid after simulcast->unicast negotiation fails
  assert_equals: expected "success" but got "failure"
Harness: the test ran to completion.
